Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

The changes in this set are rather superficial but I still have some concerns.

One, the entire purpose of this change appears to be renaming references to "SIM" to "ICC".  why "ICC"?  Technically, the SD card is an ICC as well.  Maybe we could use something more descriptive instead of jargon?

Two, the search-and-replace does not even seem to be complete.  There are still plenty of references to SIM, even in the files that are being modified here.

Patch-set: 1
Reviewer: Ficus Kirkpatrick <1000492@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1