Fix bug 1734302: Auto refresh of Contact names fails.
Need to invalidate the id->address cache in the adapter when a contact
changes. The old phone URI registered with the content resolver didn't
seem to be the right one. The observer's onChange never got called when
adding/deleting/changing a contact. I changed the uri to People.CONTENT_URI
and now onChange is getting called appropriately. Could this account for
missing presence updates?