Checkpoint work on MMS/SMS Recycler
A user can enable a setting to limit the number of saved messages
per thread. Eventually, UI will allow a user to lock a message and
prevent it from getting automatically deleted when hitting the limit
of saved messages. Work to do includes adding a saved column to the
sms and pdu tables and then implementing the UI to modify/view the
lock per message.