Fix bug 1691499: MMS constantly stopping.
What was happening was MmsSystemEventReceiver was registered for any connection
state changes in the AndroidManifest. On a connection state change, MmsSystemEventReceiver
would startup the TransactionService. The TransactionService would check its queue and if
there wasn't anything to do, would stop itself. Apparently on the MV campus, there exists
a situation that causes the phone to switch between connections, causing the CONNECTION
STATE CHANGED intent to be sent out frequently and hence, lots of starting and stopping of
the TransactionService even when it has nothing to do. With this change, the MmsSystemEventReceiver
no longer registers for that intent in the AndroidManifest. That event receiver is still
needed in the manifest because it registers for boot completed events. Meanwhile, the TransactionService
keeps another MmsSystemEventReceiver. When the TransactionService has a transaction to
process, but there's no connectivity, it enables the MmsSystemEventReceiver to
listen for connectivity changes. Whenever the TransactionService completes its work, it
unregisters the MmsSystemEventReceiver so it no longer responds to connection changes.
I added test code to TransactionService so I could simulate no connection and test that
the MmsSystemEventReceiver registers for connection changes and then appropriately starts
up the TransactionService. I left the test code in place, but it is disabled.