Display Mode 00 for CDMA Messaging
The Message Display Mode subparameter is used to indicate to the mobile
station when to display the received message.

The Message Display Mode subparameter can take on the following parameters:

  '00': Immediate Display: The mobile station is to display the received
        message as soon as possible.
  '01': Mobile default setting: The mobile station is to display the received
        message based on a pre-defined mode in the mobile station.
  '10': User Invoke: The mobile station is to display the received message
        based on the mode selected by the user.
  '11': Reserved

So when phone receives SMS with display mode 00, the message has to be displayed
immediately.  The current android code supports Display mode 00 but it does not
give the additional requirements.  User has choice to save it or discard it after
reading. If the user doesn't respond in 5 minutes, the message will be saved to
the database automatically.