Stop the smilPlayer as soon as Back or Menu buttons are pressed
Fix the usecase where a user receives a 3gp file from MMS and plays it.
After it starts playing, if the user presses a back button, the user
observes a black box on top of message menu upto size of video Attachment.
With this fix user no longer sees any black box on top of message menu.