Fix bug 1930462: Draft message disappears in Conversation List when editing a draft.
When a user creates a draft and then backs out, a thread is created for the draft. When
the user re-edits the draft, the thread is deleted after loading the conversation. When
the user backs out again, we still have a valid threadId (but no actual thread) and the
message is saved without a corresponding thread. The fix is simple. When the thread is
deleted, the threadId has to be reset in the Conversation to match. Then when the user backs
out again (or in any other code path, e.g. send), the thread will get created.