Read a long instead of an int Thread ID
Bug 1973603 is throwing an exception when adding a new conversation. Immediately
before, the code checks to see if the conversation exists, by thread id, and if
not, creates and adds the new conversation. I'm not able to duplicate the bug,
but found a problem where we're reading the threadid as an int, rather than a long.
Also added the threadid to the exception so we'll be able to look at the database.