Fix a problem with drafts
This is a fix for bug 1973603. If you go into SMS from contacts and create a draft
and in the meantime receive a message from that same contact, we'd crash in the
conversation cache trying to add the new conversation. This crash occured because
the draft conversation wasn't getting cleared out of the cache after thread was
deleted on disk. Since conversations in the cache are "equal" by recipients, the
incoming recipient matched the stale conversation in the cache. It gets even more
tricky. Bottom line: anytime a thread is deleted from the database, the equivalent
conversation has to be removed from the cache.