Add an exception to catch a bug.
Bug 1898937 found an unusual circumstance where the original message was not null,
but after splitting the message, we're getting a message count of zero. Throw an
exception and log the original message. The UI normally doesn't allow an empty
message to be sent. Need to log this to track it down.