Avoid recipient duplication in Messaging
If you try to send an SMS/MMS to a contact which has a comma in
their name field, the Messaging app will add a duplicate recipient
for each of the commas in their name.

The RecipientsEditor tokenizer only scanned directly for commas
and then looked at the spans overlapping with those regions. This
change will take into account the span end each time and adjust the
current index to be at the end of the span if necessary.

Fixes http://code.google.com/p/android/issues/detail?id=2739