MediaPlayer.prepare() fails if the audio track in a media clip does not contain valid audio data
The reason is that no display surface is set before the prepare() call in MMS application and thus the
vidoe data path does not exist. As a result, only audio data path is verified internally at the framework
layer. If there is no valid audio data, prepare() simply fails (reasonable!)

This change contains a workaround for fixing this edge case in MMS application. A long-term solution
would be to look for the metadata information from the database established by application such as
Gallery and there is no need in MMS application to duplicate the effort.

bug 1935341