Fix a MMS crasher
This is the fix for bug 1975670, though if you read that bug, you'll swear it's
talking about a different one than my fix description which follows. When a user
receives a MMS with an attachment and goes into ComposeMessageActivity to read it,
a new WorkingMessage is created in case the user types a response. Once inside CMP,
the user may click on the attachment (to see the picture, etc.). Eventually, CMP's
onStop gets called. The WorkingMessage is empty so it is discarded. Once the user is
back in CMP, they're sitting on a time bomb with a discarded WorkingMessage. If they
try and delete the thread (the original bug), they'll get an exception that the working
message has already been deleted. This simple fix is to undelete the WorkingMessage
in onRestart.