Get protected content type
When we've got protected content, we must reach into
the drm object in order to get the actual content type.
In addition, the previous code was independently setting
the type and then the data of the intent. Both setType
and setData wipe out the other guy (by design). They both
have to be set at the same time for this to work.