SMS to own number creates two threads and crashes
Send an SMS from a test phone to itself (say the number is 9243176013).
A thread gets created.  But when this SMS gets received (of course from
the test phone itself), the number gets received as '09243176013'. And
one more thread is created. But in the conversation list, 2 threads are
shown with the same number (9243176013). When it is tried to add this
new conversation thread to the cache, it throws an Exception which is
not caught.

From this time onwards, everytime we launch Messaging, we keep getting
the Uncaught Exception, and the application closes.

To fix, in Conversation.java handled the exception by catching it from
wherever this function is called.  In files Conversation.java and
SmsReceiverService.java did changes to consider 9243176013 and 09243176013
as a same 'from address', so that only one thread id gets created.