Don't show Anonymous as the MMS sender
When the MMS app receives a message of type MESSAGE_TYPE_NOTIFICATION_IND = 0x82,
it tries to extract the sender out of the pdu headers. In the case of a particular
platform/provider, the sender header is not set. Until now, we'd fill in the sender
with "Anonymous". Eventually, another pdu arrives with the headers filled out and
the Anonymous is replaced with the real sender. The flashing is annoying, so just
let that field be empty.