Send option not enabled after composing message with slides
Issue: Compose a slideshow (with 2 or more slides), and with some valid
destination address. Say done.

It is seen that the 'Send' button is not enabled. User has no way to send
a message with slides directly.  It has to be saved in draft and opened.
And only then the 'Send' button is enabled.

Reason: When the done button is pressed, 'Send' button of the Attachment
editor is not enabled.  Fixing this by updating the Send button.