Cache not getting refreshed when deleted thread from ComposeMessage screen
When a thread is deleted from ComposeMessageActivity it is not getting deleted from cache.
And sometimes the deleted thread's address is getting displayed while forwarding other SMS.
Although cache is refreshed when a thread is deleted from ConversationList, when deleted
from ComposeMessage, it is not refreshed. Adding code to refresh the cache on a thread deletion...