One-time storage limit check
When running the Messaging app for the first time, specifically the
ConversationList, check to see if we can auto-enable the storage limits
setting. If any of the conversation threads have more messages than
the storage limit settings, show an activity that will tell the user how
to turn on the storage limits. Otherwise, just enable the setting.