Contacts with commas
If a contact name contains a comma (i.e. "Flintstone, Fred) the messaging app would
consider that two contacts because it uses comma as a delimiter. The app would end
up sending the same message to that contact twice. Don't add dupes to the final
list of numbers to send to.