MMS-IOT Single Segment WAP duplicate check always succeeds
The content location is used to detect duplicate MMS WAP
pushes. This is a problem with single segment WAP pushes
since these always have the same Content Location until the
Transaction ID is appended to the Content Location URL.
User was only able to receive a single MMS message.
Solution is to append the Transaction ID to the
Content Location much earlier in the process (before the
duplicate check). The URL is persisted in the database,
with TransactionID attached, making it unnecessary to check
the feature enable flag in any subsequent code.