Compose Message "To" field allows adding sequence of characters as contact
When we select this option, this sequence of characters (which is neither a
valid email address nor a phone number), gets stored in contact with type MOBILE.

Fix: In function canAddToContacts, adding a check to validate whether the contact
to be added is a valid email address or phone number. If not, user will not be shown
with the option "Add to Contacts".