Remove Manage SIM pref
Dynamically remove the "manage sim" pref when the device doesn't
have a SIM. Also, simplify the existing code. There was no need to
create a Preference subclass just to handle clicking on the preference.
The actual fix has been short-circuited until the function used to
check for the presence of the SIM works correctly.

Change-Id: I465256f675a059459000a2c4832e97259c24efc0