Fix send order
When SMS messages can't be sent, they're queued up. When messaging
can continue (back in range or out of airplane mode) the messages
were pulled off the queue in the newest-to-oldest order, resulting in
them getting sent in the wrong order. With this change, messages are
send in the same order that they were composed.

Change-Id: Id7a190dc27ef975c9553fd33210c58bba07ef0dd