Resize images when they exceed the MMS payload size
Code was in place to resize images when they exceeded the maximum
MMS dimensions. But the case where the byte size of an image was exceeding the
MMS payload just resulted in an error message to the user. This is a problem
with high resolution images taken by the devices camera. These images fit the
MMS dimensions but routinely exceed the MMS byte size.

Extra fix for GIF images. The GIF decoder does not honor the scaled decode
parameter. Code now calls the Bitmap scale function when this happens.

    1. Change the getResizedImageData function in UriImage.java to account for
       byte size as well. If the resulting jpg exceeds the byte size attempt to
       reduce the quality on the compressor first. But not below 50%. Should
       this step not work then rescale the image to a lower resolution and try
       again. Do this up to 4 times.
    2. Change resizeImageAsync in MessageUtil.java to pass a maximum bytes size.
       Account for a MMS overhead of 5000 bytes in this case. So the maximum
       image size will be msConfig.getMaxMessageSize() ? 5000.
    3. Update 3 places to call resizeImageAsync instead of showing an error
       message to the user.

Change-Id: I33427fe79c42677e24acf0390ae834d58205c8f1