Confirm delete locked messages
Allow a user to delete a locked message, but show a different
dialog alert so the user will know they're specifically deleting
a locked message. Also, fix deleting a thread. Use the same code
for deleting a thread as used in ConversationList where we check
to see if the thread contains locked messages and if so, present
a different UI.