Force-close when adding picture to a slide having video
Issue: Add video to a slide. There is a button 'Add picture'. Click it, and select a picture.
Messaging force closes.

Rootcause: It is obvious that a slide cannot have a video as well as picture. When doing so,
the function internalAdd() gets executed and throws IllegalStateException, which is not caught
anywhere.

Fix: When a slide has a video already, and user clicks 'Add picture', show him toast saying
'video and picture cannot added to one slide' and return.

Bug: 2111946
Change-Id: Iad57c2f664569bbf20db00c477d1e8f0b22664f9