Better handle filenames
When getting an image uri that pointed to a file, the code didn't
handle filenames with spaces very well. Furthermore, if the code couldn't
parse the filename extension, it would throw an exception and crash.
Give another shot at finding the extension and if that fails, don't crash,
but show a "picture couldn't be attached" toast.

Change-Id: I5a2fcf6cc16984d81a3b8d5a7ef19367e16dd839