Fix logging
There's a situation where the incoming intent has a '%' in one of
the extras. When logging is enabled via setprop the log()
function called and trying to format the string causes an exception because
there are not format parameters. It turns out in CMA, we're only
passing strings, but no parameters, so simplify the logging function
to only take a string. We could make this change in other Mms files,
but for now, this one is the only one with a report of a crash. Users
won't ever see this because logging is off by default.