Added a new option in the settings to specify the number of SMS messages at which the message will be converted to an MMS. This fixes #2318.
Note that carriers or manufacturers might want to impose fixed limits on this setting or eliminate it altogether. This can be done by either modifying the string array with the possible settings, or removing the option from preferences.xml entirely. There is also a hard-coded private int, MAX_MMS_THRESHOLD, that specifies the upper limit for this setting. The user will not be able to exceed this value, not even by hacking the preferences file on the device.

See also this thread on android-platform: http://groups.google.com/group/android-platform/browse_frm/thread/b7cb1a54528a4c1b