Update patch set 1
Romain, you're right. My question is - how do you solve the issue? You could a) increase the limit internally (I suggested that, but people advised that carriers might not go for that), b) tell people that this is the way it is and close #2318, or c) add a setting for it. I chose c) to address the issue and to keep carriers/manufacturers in control, but I'm interested in any alternative approach.

(This is why I was really hoping for somebody to respond on android-platform BEFORE I committed this change.)

Patch-set: 1