Fix draft bug
Fix a special case of working on a draft and then getting a new intent for that same
draft. Recognize we've got the same conversation rather than trying to save the draft
and reload the same draft. Add more logging to Conversation and WorkingMessage and move
the common logging code into LogTag.

Change-Id: I1388967f3f2e7520e08730036243f958574a6001