Don't use auto-requery
The default behavior with CursorAdapter is to do an automatic requery when
the content has changed. This happens on the UI thread and can cause ANR's.
Handle the content change notification ourselves and do a new query asychronously.
Update ConversationList to handle content changes and do a background query,
rather than relying on the auto cursor requery. The rarely used ManageSimMessages
can still do an auto-requery, but it's less likely because we don't write to nor
change messages on the SIM. Also, a different content observer is already registered
to notice SIM changes and update the message list in the background.

Change-Id: I42ac518b315ea55ff6edbe0a912bf60c7f868ea5