Don't change focus
When opening the keyboard, don't change the focus. The line I'm removing
must have been put in to solve some problem, but it causes other problems.
For instance, if you're editing a recipient in portrait using the soft-keyboard
and decide it'd be easier to use the hard keyboard, the focus will switch out
from under the user from the recipients editor to the message editor. With this
change, the focus will remain in the field the user was typing in.

Change-Id: I84e31b9685aae936579d1b728750ed5edc0c8c60