Change dialog to toast
If we get an exception while adding a picture because the picture is
too large, a function is called to downscale the picture. If that process
take over a second, a "Compressing..." dialog is shown. If the user opens
and closes the keybaord during this period, the dialog is left dangling
or attached to the wrong window and this can cause several different crashes.
This checkin replaces the dialog with a toast that seems more immune to
the same steps. In other words, the toast closes itself, rather than us
having to call the dialog.dismiss function, which was problematic.
Also, fix a merge change on removing commas from recipient names.

Change-Id: I4a82fd6f832d935915583efb860a6c0382b3a2df