Handle out-of-memory while decoding image
When adding images to a mms message we attempt to resize an image, if
necessary, to keep the message within its maximum message length. In
some cases, the first crack at decoding the image can throw an
out-of-memory exception. With this change, we catch that exception and
try decoding again at a smaller scale.

Change-Id: Id8ebcab4c407f316dd850640ab8f058845fb7054