bug 2169583: fix a thread mixup problem caused by empty draft message having an unaccounted for thread id.
- there is a situation where if the user tries to attach an image in an empty draft message, we'd allocate
a thread id (creating one in the db), but immediately deleting it because the draft message is empty.
However, when we deleted the empty draft message, we didn't invalidate the in-memory thread id from the
initial allocation. At this point if a new message arrives and takes the same thread id, it will cause
the CMA to requery and pull in the new message to the draft message thread incorrectly (because both
have the same thread id).

Change-Id: I7981eb1ec1f4ba1e2650325093c2b60ac4324b94