Fix message item click
When downloading a Mms video, the message item shows a play button.
Because there are various levels of nested click handlers, only the
button responds to clicks. That made it impossible for users to click
links in the text portion of the message. Turn on the click handler for
the item so both the item and the button can respond to clicks.

Change-Id: I76f5f5fb832a1bb40b9eeff2b22d0f21ee88371f

Change-Id: I5472672da2119cf01c2ab306dfa5111c93ddf478