Fix contact header updating
When a user added a recipient and tapped the message area to start
composing, the code would ask for the contact. Since we didn't have
the contact in the cache, we'd fire off an async request to load the
contact info. Meanwhile, ComposeMessageActivity would use the contact's
name, which in this case was just the number, to display in the contact
header widget. With this change, we register for changes on all the
contacts. If a contact changes, we rebuild the contact header widget.
Also removed obsolete presence code.