Allow full-length video replacement
When replacing an attachment with a different attachment, take into
consideration the current size of the slide and subtract that out
of the max message length. Right now, we're only doing that for video
as it is currently the only attachment where we can specify the limit
in advance. The technology is now there, however, to do the same with
captured audio. Fixes bug 2227514.