bug 2236527: need to scrub the phone number before sending a MMS message.
- add MMS address validity checking and scrubbing. If the phone number used to send a MMS message
contains invalid characters (non digits and non-valid separators), we now warn the user when he tries
to send the message. The prior validity checking code only checks against GSM SMS address, so it
would miss all the invalid chars for MMS address.

- assuming the MMS address is ok but contains separators that are not valid MMS address chars (as
far as the network is concerned), we scrub the number when creating the SendReq data structure for
MMS message.

- flag a few places where things are still not correct, but too risky to fix for MR1.

Change-Id: Ia80a2736f955b94a8cac51ffeb06617a95580ab8