Speedup contact name resolution
The first time Messaging is launched, a background thread loads
the conversations threads, which loads the contact info cache. While
this is happening, the ConversationList activity starts and the
first thing it does is clear the contact cache that is being built
in the background. Don't clear the cache if it's currently being
built. Fixes bug 2167799.

Change-Id: Id8a9345caaf7905d1f4bcc18ee45d11cb860d293