Fix missing From field
In rare cases, a Mms pdu header will not contain a "from" address.
Rather than leaving the from address blank, use a slower, but more
reliable method of getting the from address from the "addr" table.
This is the technique the Mms notification system uses to get the address
displayed in the notification bar when a new message is received.
Fixed bug 2241939.

Change-Id: If932d4f1c24f46f3c3906a7e2658d7f67d4a2cfa