Fix ANR deadlock
When the Mms app starts up, Conversation starts a background thread
to load all the conversations into a cache. It had a lock on the cache.
Meanwhile, the ComposeMessageActivity started up and tried to load
its conversation, but was blocked and eventually ANR'd. Make the
sync lock cover a much smaller scope and don't hold the lock during
the slow operation of looking up the contact info. Also, remove the
unused SuggestionsProvider. It was causing the Mms app to spin up
needlessly. Fixes bug 2220565.

Change-Id: I0f8260338a5368dc4b8ef11249361ca0716a2887