Only register for service changes when needed
The SmsReceiverService registered for SERVICE_STATE changes and this
was causing the Messaging app to be needlessly started up on every
SERVICE_STATE change. With this fix, we only register for that event
when we fail to send a SMS message because of no radio sig, etc.
Once we successfully send the queued message, we'll unregister the
event filter.

Change-Id: I12cbe0cb9a72bbf628bf9a4929680519e85057a2