Fix ANR from calling Contact.get()
The Contact.get() API allows the caller to specify whether the operation
can block (looking up contact info in the DB) or not. There are about 10
places where various callers were passing "true" to canBlock, all coming
from the UI thread. We've previously fixed Contact.get() to always return
a Contact, with the number filled in, even if we don't have the contact
in the cache. With this change, never block the UI thread getting a contact.
The worst side-effect should be showing a number, instead of a name.
Fixes bug 2265631.