Fix auto-requery ANR's
There are a couple of bug reports of ANR's from CursorAdapter doing
a requery on the UI thread. There are two off-the-beaten-path places
where the CursorAdapter was getting set with auto-requery == true
by default. Also, the recipients editor adapter had requery == true.
With this change, there should be no auto-requerying in mms.