use a hashtable to speedup contacts cache lookup
The strategy is now as follows:  for any given phone number take
the last 5 digits and use that as a key in a hashmap.  Map to a
list of Contacts all of whom share that same key.  Typically that
list will be very small.

Each "get" still requires at least one call to PhoneNumberUtils.compare.
We could eliminate that by introducing another hashtable which maps
from phone number to Contact object.  In that case the existing mapping
function would only be used as a fallback when we didn't find anything
in the direct hash.

I recommend reviewing this change by looking at the new implementation of
"get".  Trying to look at "diffs" won't work very well.