Cleanup toast handler
This is for bug 2369881. The bug recommended removing the handler.
When I tried that (by calling the toast directly) the toast no longer
showed up. I stepped through the toast's show() method, but didn't see
anything wrong. I decided to follow the model of Activity's runOnUiThread()
and make the code a little more obvious.

Change-Id: Ice20fe9e8dff121ec2a54432bbccdab826537633