Update ConversationList backgrounds
Now going with a light background theme, instead of a dark one.
Take the opportunity to rename several files to names that are
more standard across Android apps. This will make it easier to
find the listitems that belong to the conversation list. The old
names always trip up those browsing the sources for the first, second,
and tenth times. Fixes bug 2406873.