Remove extra ScrollView
A ScrollView wrapping the attachment editor and message field
was causing problems. After sending an MMS and hiding the attachment
editor, the message editor would be scroll off-screen. Tens of
seconds later it would scroll back into view. There's no need for
this scroll view in the first place. Bug 2406769