Fix double display of outgoing messages
Prior to this fix, sending an SMS message would put the message
in the SMS table marked as pending. As the message was pulled off the queue
to be sent, a duplicate message was added to the SMS table and marked
as "outbox". Then the pending message was deleted out of the same table.
Meanwhile, the UI responded to these DB changes and momentarily flashed
two copies of the outgoing message in the ComposeMessageActivity. With
this change, we only keep a single copy of the message in the SMS table and
simple change the column value from pending to outbox. The result is a
lot less DB churn and a much smoother UI when sending an SMS. Tested with
short SMS messages and long, multi-part messages.

Change-Id: Iaad4f7a9732d79786b50aa19428f51bfec002d4a