Don't resize image on attach
Previously, image attachments would get resized as soon as they
were added to a message. This caused a problem in that the first
picture would eat up as much space as it could, the second picture
would get the scraps, and on down the line. Now we don't resize
attachments until the message is actually sent and at that point,
give each attachment equal space to resize. However, there was one
unneeded resize going on when individually attaching a picture.
The picture would get resized down to the carrier's max h & v
resolution. With this change, we skip that extra resize and the
max h & v resolution gets taken into account on the final resize
at send time.