Make room for compose field
Recent changes to the compose area have made it slightly taller to
allow for the send button to show two lines of text. When the
on-screen keyboard slides up, the compose area gets squished. Have
to compensate by making the buttons in the various attachment editors
5 pixels shorter. Also, wrap the slideshow area in a scrollview to
handle different screen sizes. Bug 2440189.

Change-Id: I1c6b6496d628aa3104ade527d0bf73e29cc59c98