Fix draft corner case
When the message being edited is the only message in the conversation, the delete
message gets deleted before being edited. The trigger "delete_obsolete_threads_pdu"
sees that a thread contains no messages and silently deletes the thread.
Meanwhile, the mConversation object still holds onto the old thread_id.
The code now notices that situation and clears out the thread_id.
Later on, when Conversation.ensureThreadId() is called, we'll
create a new thread if necessary. Bug 2166889

Change-Id: Id5d4b810dbef87f67e8aa9084faf4b7bd8557972