Bug 2030420: fix SMS/MMS statusbar notification.
- don't show statusbar notification for sms/mms based on whether there are unread messages or not. Instead, add a "seen" column for the SMS and Mpdu (mms) table. Each SMS and MMS message will now have a "seen" flag in the db. If there are unseen messages, we'll show the statusbar notification. Once the user reads a thread, we'll mark a thread (and all messages in the thread) read and seen. If the user goes to the conversation list activity, we'll mark all messages "seen", but won't touch the read flag for the messages.

So now the user can have many unread messages, but still have seen all the messages, thus won't get notified repeatedly just because there are unread messages.

we'll mark a message "seen" flag to 0 for incoming SMS and MMS (pdu) messages.

to-do: mark all messages as seen when the user hits the clear notification button in the window shade.