Avoid unnecessary DB updates
Making database updates at the same time we're trying to read the list of
messages in a conversation causes any further queries to block until the updates
are complete. This results in taking a long time to display a new
message when the user selects the sms notification. With this change,
we block marking a thread as "read" until we've completed the query for
a conversation's list of messages. Also, every delete causes several
DB updates inside the provider. We can avoid one of these deletes that causes
so many updates if the thread has no drafts. Now responding to a new message
notification only results in two updates (one to the sms table and one to the
pdu table) and that happens after the messages are loaded.
Bug 2467382.

Change-Id: I2a78023e3ba7dee72c110d2023f4c79b1b816565