Don't delete draft unless we need to
Don't even look for a draft unless the conversation says we've got a draft.
Don't delete the draft unless the draft is the only message in the
conversation. Excess drafts will get deleted at send time (or in various
other places where we save the draft or discard). Fix one of the
Conversation constructor's to use the cache. Otherwise, ConversationList
activity was updating all the conversations, but all the other activities
were retrieving old conversations from the cache. Bugs 2483022 and 2329323.