Fix empty draft
When an attachment is added, the working message is bumped up to a
temporary Mms state. Once the attachment is added, the various
flags in working message will be updated and that temporary state
removed. However, if the user backs out of adding an attachment,
the working message will be left in the fake Mms state and on
exit, will cause an empty draft to be saved. Now we clear that
fake Mms state when we return from adding a draft, successful or not.
Bug 2493456.

Change-Id: I596efbee0bcd54c30b1de4f4aedf1b9c6919f947