Fix race condition
After deleting a thread in ConversationList, the onDeleteComplete function
would call Conversation.init() to update the threads and make the
conversation cache match what was on disk, then call
startAsyncQuery() to get the new state of the threads. Then it called
onContentChange(), which was a mistake. onContentChange() caused the
list to redraw based on the old cursor. The code would reinsert
the deleted conversation back into the cache as a result of doing
a Conversation.get. Later on when looking up a thread by thread id,
the cache could easily return the deleted conversation from the cache.
With that onContentChange() removed, the async query takes place
and the conversation list gets rebuilt with the new cursor.
Fixes bug 2492283.

Change-Id: I901e774dd341a3ed1a39f9a47e8026555f853566