Fix an odd startup bug
Bug 2490108 describes a situation where starting a new messaging
conversation from Contacts brings up the compose message activity in a
state where the recipients editor is not editable, even though it's a
new conversation. The bug also describes the normal case where the editor
does come up editable. The difference is that one goes through onCreate
and the other through onNewIntent. We need to clear a flag that says
whether a message has been sent or not so the recipient editor will be
displayed properly in onQueryComplete.

Change-Id: Idaacea20c277dc39e9d258fd73354c915119c2ed