Fix cursor out of range
There's a scenario where long-pressing on a message while a
message receipt comes in causes a cursor change. By the time the
onMenuItemClick handler gets called, the cursor is invalid.
Code would test for a valid cursor, but unfortunately, earlier
code would try and read values from that cursor. With this fix,
we test for a valid cursor before doing any operations. Bug 2511080.

Change-Id: Ib689ba9cf58ca7dc552c2c258b8043c4be59c0c6