Don't edit the wrong item
If a new message arrives or the underlying DB changes at the same
moment the user taps the error icon to edit and resend a message,
the editor can come up showing the wrong message. When the DB
gets updated, MessageListAdapter.notifyDataSetChanged() clears the
mMessageItemCache. At the same time, editing a message with an
error calls mMessageListItemhandler.handleMessage, which calls getMessageItem
which would try and get the messageItem out of the cache. Since the
cache is empty, it would then create a new messageItem out of whatever
the cursor was pointing to! With this change, the code in this path
will not create that bogus item and will silently do nothing. This
same function is called from the menu code and I argue the menu code
shouldn't be creating items either, but I didn't make that change in
this CL. Bug 2372633

Change-Id: I009a4b561e5d7a81cb45724fdf3041e9e413e09a