Alert on non-UI thread
Now that we're doing more stuff off of the UI thread, catch a case
where we were trying to put up an error dialog while attaching a video.
The function is called from many places. Make sure the guts of the
function that show a toast or dialog run on the UI thread. Bug 2530897

Change-Id: I8671f43b29ccfd44f48b9f224188fa9ac6648370