Fix sms and mms draft confusion
When saving an mms draft, any older sms draft associated with that
same thread is deleted. That was the intent. However, the code for
getting the thread_id was in a separate thread. The code that deleted
the old sms draft was outside that thread and the conversation's
thread_id was still zero. Very simple fix. Bug 2534828.

Change-Id: Ice72770bdd506840883afd9f81a66d3447b3b764