Update patch set 1
Patch Set 1: No score

The only reason we added this check here is because an ACK needs to be sent out even though the message is dropped. Following the code flow, the ACK was sent in framework layer after the Intent is sent to the App.So we chose to drop the message here.

In retrospect it does make more sense to add the check in GsmSMSDispatcher.java.since the requirement is specific to Gsm. Will analyze a bit more if the change can be done in framework such that the message is dropped but an ACK is sent out and upload a new patch if possible.

Patch-set: 1
Reviewer: Naveen K <1003413@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0