don't crash trying to show the delivery report if it doesn't exist
Apparently you can click on the error icon to go back into compose
message activity with the original text and original address
prepopulated -- so you can fix it up.  But for reasons that aren't
clear (it's not clear if it's even intentional) we continue to also
show the failed message item with the failure icon.  If you long
press on that item we include a menu item for showing the delivery
report.  Clicking on that fails to find the item in the sms or mms
table and we crash.

This change makes it so that we're more resilient to this state.
In other words we don't crash.  The dialog which shows the delivery
report is somewhat ugly in this case but acceptible in my opinion.

Change-Id: Ie34c32c83982ff0d506e0102813370e1e7c49a12